web.py
from aiohttp import web
import logging

logger = logging.getLogger('WEB')

async def hello(request):
    return web.Response(text="Hello, world")

app = web.Application()
app.add_routes([web.get('/', hello)])

async def run():
    logger.info("Website starting...")
    runner = web.AppRunner(app)
    await runner.setup()
    site = web.TCPSite(runner, '0.0.0.0', 8080)
    await site.start()