requirements.txt
requests
discord.py
flask
userutils
dateparser==0.7.2
